workshop/workshop-basic-iot/data-publisher/requirements.txt
paho-mqtt
python-dotenv
pandas